refactor: rename base package to com.clienthub and improve class naming
