+++ backend/pom.xml
@@ -8,7 +8,7 @@
 		<version>3.1.2</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>com.havefunwith</groupId>
+	<groupId>com.clienthub</groupId>
 	<artifactId>client-management-api</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>client-management-api</name>
