+++ backend/src/main/java/com/clienthub/customer/CustomerCreateRequest.java
@@ -1,0 +1,8 @@
+package com.clienthub.customer;
+
+public record CustomerCreateRequest(
+        String name,
+        String email,
+        Integer age
+) {
+}
