+++ backend/src/main/java/com/clienthub/customer/CustomerInMemoryRepository.java
@@ -1,0 +1,67 @@
+package com.clienthub.customer;
+
+import org.springframework.stereotype.Repository;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Optional;
+
+@Repository("list") // naming bean
+public class CustomerInMemoryRepository implements CustomerDataAccess {
+
+    private static List<Customer> customers;
+
+    static {
+        customers = new ArrayList<>();
+
+        Customer john = new Customer((long) 1.0, "John", 24, "john@email.com");
+        Customer keyla = new Customer((long) 2.0, "Keyla", 28, "keila@email.com");
+
+        customers.add(john);
+        customers.add(keyla);
+    }
+
+    @Override
+    public List<Customer> selectAllCustomers() {
+        return customers;
+    }
+
+    @Override
+    public Optional<Customer> selectCustomerById(Long id) {
+        return customers.stream()
+                .filter(c -> c.getId().equals(id))
+                .findFirst();
+    }
+
+    @Override
+    public Long insertCustomer(Customer customer) {
+        customers.add(customer);
+        return customer.getId();
+    }
+
+    @Override
+    public boolean existsPersonWithEmail(String email) {
+        return customers.stream()
+                .anyMatch(customer -> customer.getEmail().equals(email));
+    }
+
+    @Override
+    public boolean existsPersonById(Long id) {
+        return customers.stream()
+                .anyMatch(customer -> customer.getId().equals(id));
+    }
+
+    @Override
+    public void deleteCustomer(Long id) {
+        customers.stream()
+                .filter(customer -> !customer.getId().equals(id))
+                .findFirst()
+                .ifPresent(c -> customers.remove(c));
+    }
+
+    @Override
+    public void updateCustomer(Customer customer) {
+        customers.add(customer);
+    }
+
+}
