+++ backend/src/main/java/com/clienthub/customer/CustomerJpaRepository.java
@@ -1,0 +1,56 @@
+package com.clienthub.customer;
+
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+import java.util.Optional;
+
+/*
+    Data Access Service Layer
+ */
+@Repository("jpa")
+public class CustomerJpaRepository implements CustomerDataAccess {
+
+    private final CustomerRepository customerRepository;
+
+    public CustomerJpaRepository(CustomerRepository customerRepository) {
+        this.customerRepository = customerRepository;
+    }
+
+    @Override
+    public List<Customer> selectAllCustomers() {
+        return customerRepository.findAll();
+    }
+
+    @Override
+    public Optional<Customer> selectCustomerById(Long id) {
+        return customerRepository.findById(id);
+    }
+
+    @Override
+    public Long insertCustomer(Customer customer) {
+        customerRepository.save(customer);
+        return customer.getId();
+    }
+
+    @Override
+    public boolean existsPersonWithEmail(String email) {
+        return customerRepository.existsCustomerByEmail(email);
+    }
+
+    @Override
+    public boolean existsPersonById(Long id) {
+        return customerRepository.existsCustomerById(id);
+    }
+
+    @Override
+    public void deleteCustomer(Long id) {
+        customerRepository.deleteById(id);
+    }
+
+    @Override
+    public void updateCustomer(Customer customer) {
+        customerRepository.save(customer);
+    }
+
+}
