+++ backend/src/main/java/com/clienthub/customer/CustomerRepository.java
@@ -1,0 +1,8 @@
+package com.clienthub.customer;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
+    boolean existsCustomerByEmail(String email);
+    boolean existsCustomerById(Long id);
+}
