+++ backend/src/main/java/com/clienthub/customer/CustomerUpdateRequest.java
@@ -1,0 +1,8 @@
+package com.clienthub.customer;
+
+public record CustomerUpdateRequest(
+        String name,
+        String email,
+        Integer age
+) {
+}
