+++ backend/src/main/java/com/clienthub/exception/DuplicatedResourceException.java
@@ -1,0 +1,13 @@
+package com.clienthub.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.CONFLICT)
+public class DuplicatedResourceException extends RuntimeException {
+
+    public DuplicatedResourceException(String message) {
+        super(message);
+    }
+
+}
