+++ backend/src/main/java/com/clienthub/exception/ResourceNotFoundException.java
@@ -1,0 +1,13 @@
+package com.clienthub.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.NOT_FOUND)
+public class ResourceNotFoundException extends RuntimeException {
+
+    public ResourceNotFoundException(String message) {
+        super(message);
+    }
+
+}
