+++ backend/src/main/java/com/clienthub/singleton/AppConfig.java
@@ -1,0 +1,14 @@
+package com.clienthub.singleton;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class AppConfig {
+
+    @Bean
+    public Product createProduct() {
+        return new Product("Shoes", 50.0);
+    }
+
+}
