--- backend/src/main/java/com/havefunwith/customer/Customer.java
@@ -1,114 +1,0 @@
-package com.havefunwith.customer;
-
-import jakarta.persistence.*;
-
-import java.util.Objects;
-
-@Entity
-@Table(
-        name = "customer",
-        uniqueConstraints = {
-                @UniqueConstraint(
-                        name = "customer_email_unique",
-                        columnNames = "email"
-                )
-        }
-)
-public class Customer {
-
-    @Id
-    @SequenceGenerator(
-            name = "customer_id_seq",
-            sequenceName = "customer_id_seq",
-            allocationSize = 1 // will match increment size in the entity mapping to the associated database sequence
-    )
-    @GeneratedValue(
-            strategy = GenerationType.SEQUENCE,
-            generator = "customer_id_seq"
-    )
-    private Long id;
-    @Column(
-            nullable = false
-    )
-    private String name;
-    @Column(
-            nullable = false
-    )
-    private Integer age;
-    @Column(
-            nullable = false
-//            unique = true // it's being defined at @Table annotation
-    )
-    private String email;
-
-    public Customer() {}
-
-    public Customer(String name, Integer age, String email) {
-        this.name = name;
-        this.age = age;
-        this.email = email;
-    }
-
-    public Customer(Long id, String name, Integer age, String email) {
-        this.id = id;
-        this.name = name;
-        this.age = age;
-        this.email = email;
-    }
-
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public Integer getAge() {
-        return age;
-    }
-
-    public void setAge(Integer age) {
-        this.age = age;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Customer customer = (Customer) o;
-        return Objects.equals(id, customer.id) && Objects.equals(name, customer.name) && Objects.equals(age, customer.age) && Objects.equals(email, customer.email);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, name, age, email);
-    }
-
-    @Override
-    public String toString() {
-        return "Customer{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                ", age=" + age +
-                ", email='" + email + '\'' +
-                '}';
-    }
-}
