--- backend/src/main/java/com/havefunwith/customer/CustomerController.java
@@ -1,60 +1,0 @@
-package com.havefunwith.customer;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
-
-import java.net.URI;
-
-import java.util.List;
-
-// Customer REST controller - handles all incoming HTTP requests for customer operations
-// Endpoint: api/v1/customers
-@RestController
-@RequestMapping("api/v1/customers")
-public class CustomerController {
-
-    private final CustomerService customerService;
-
-    // dependency injection
-    public CustomerController(CustomerService customerService) {
-        this.customerService = customerService;
-    }
-
-    // @RequestMapping(value = "/api/v1/customers", method = RequestMethod.GET)
-    @GetMapping
-    public ResponseEntity<List<Customer>> getCustomers() {
-        return ResponseEntity.ok(customerService.getAllCustomers());
-    }
-
-    @GetMapping("{customerId}")
-    public ResponseEntity<Customer> getCustomer(@PathVariable("customerId") Long customerId) {
-        return ResponseEntity.ok(customerService.getCustomer(customerId));
-    }
-
-    @PostMapping
-    public ResponseEntity<Void> registerCustomer(@RequestBody CustomerRegistrationRequest request) {
-        Long customerId = customerService.addCustomer(request);
-        URI location = ServletUriComponentsBuilder
-        .fromCurrentRequest()
-        .path("/{id}")
-        .buildAndExpand(customerId)
-        .toUri();
-        return ResponseEntity.created(location).build();
-    }
-
-    @DeleteMapping("{customerId}")
-    public ResponseEntity<Void> deleteCustomer(@PathVariable Long customerId) {
-        customerService.deleteCustomer(customerId);
-        return ResponseEntity.noContent().build(); // RESTful No content 204 
-    }
-
-    @PutMapping("{customerId}")
-    public ResponseEntity<Void> updateCustomer(
-            @PathVariable("customerId") Long customerId,
-            @RequestBody CustomerUpdateRequest request) {
-        customerService.updateCustomer(customerId, request);
-        return ResponseEntity.noContent().build(); // RESTful No content 204
-    }
-
-}
