--- backend/src/main/java/com/havefunwith/customer/CustomerJPADataAccessService.java
@@ -1,56 +1,0 @@
-package com.havefunwith.customer;
-
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.Optional;
-
-/*
-    Data Access Service Layer
- */
-@Repository("jpa")
-public class CustomerJPADataAccessService implements CustomerDao {
-
-    private final CustomerRepository customerRepository;
-
-    public CustomerJPADataAccessService(CustomerRepository customerRepository) {
-        this.customerRepository = customerRepository;
-    }
-
-    @Override
-    public List<Customer> selectAllCustomers() {
-        return customerRepository.findAll();
-    }
-
-    @Override
-    public Optional<Customer> selectCustomerById(Long id) {
-        return customerRepository.findById(id);
-    }
-
-    @Override
-    public Long insertCustomer(Customer customer) {
-        customerRepository.save(customer);
-        return customer.getId();
-    }
-
-    @Override
-    public boolean existsPersonWithEmail(String email) {
-        return customerRepository.existsCustomerByEmail(email);
-    }
-
-    @Override
-    public boolean existsPersonById(Long id) {
-        return customerRepository.existsCustomerById(id);
-    }
-
-    @Override
-    public void deleteCustomer(Long id) {
-        customerRepository.deleteById(id);
-    }
-
-    @Override
-    public void updateCustomer(Customer customer) {
-        customerRepository.save(customer);
-    }
-
-}
