--- backend/src/main/java/com/havefunwith/customer/CustomerRegistrationRequest.java
@@ -1,8 +1,0 @@
-package com.havefunwith.customer;
-
-public record CustomerRegistrationRequest(
-        String name,
-        String email,
-        Integer age
-) {
-}
