--- backend/src/main/java/com/havefunwith/customer/CustomerRowMapper.java
@@ -1,21 +1,0 @@
-package com.havefunwith.customer;
-
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.stereotype.Component;
-
-import java.sql.ResultSet;
-import java.sql.SQLException;
-
-@Component
-public class CustomerRowMapper implements RowMapper<Customer> {
-
-    @Override
-    public Customer mapRow(ResultSet rs, int rowNum) throws SQLException {
-        return new Customer(
-                rs.getLong("id"),
-                rs.getString("name"),
-                rs.getInt("age"),
-                rs.getString("email")
-        );
-    }
-}
