--- backend/src/main/java/com/havefunwith/customer/CustomerUpdateRequest.java
@@ -1,8 +1,0 @@
-package com.havefunwith.customer;
-
-public record CustomerUpdateRequest(
-        String name,
-        String email,
-        Integer age
-) {
-}
