--- backend/src/main/java/com/havefunwith/exception/DuplicatedResourceException.java
@@ -1,13 +1,0 @@
-package com.havefunwith.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.CONFLICT)
-public class DuplicatedResourceException extends RuntimeException {
-
-    public DuplicatedResourceException(String message) {
-        super(message);
-    }
-
-}
