--- backend/src/main/java/com/havefunwith/exception/ResourceNotChangedException.java
@@ -1,12 +1,0 @@
-package com.havefunwith.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.NOT_MODIFIED)
-public class ResourceNotChangedException extends RuntimeException {
-
-    public ResourceNotChangedException(String message) {
-        super(message);
-    }
-}
