--- backend/src/main/java/com/havefunwith/exception/ResourceNotFoundException.java
@@ -1,13 +1,0 @@
-package com.havefunwith.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.NOT_FOUND)
-public class ResourceNotFoundException extends RuntimeException {
-
-    public ResourceNotFoundException(String message) {
-        super(message);
-    }
-
-}
