--- backend/src/main/java/com/havefunwith/singleton/Product.java
@@ -1,38 +1,0 @@
-package com.havefunwith.singleton;
-
-public class Product {
-
-    private String name;
-    private Double price;
-
-    public Product() {}
-
-    public Product(String name, Double price) {
-        this.name = name;
-        this.price = price;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public void setPrice(Double price) {
-        this.price = price;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public Double getPrice() {
-        return price;
-    }
-
-    @Override
-    public String toString() {
-        return "Product{" +
-                "name='" + name + '\'' +
-                ", price=" + price +
-                '}';
-    }
-}
