+++ backend/src/test/java/com/clienthub/TestcontainersTest.java
@@ -1,0 +1,15 @@
+package com.clienthub;
+
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class TestcontainersTest extends AbstractTestcontainers {
+
+    @Test
+    void canStartPostgresDB() {
+        assertThat(postgreSQLContainer.isRunning()).isTrue();
+        assertThat(postgreSQLContainer.isCreated()).isTrue();
+    }
+
+}
