--- backend/src/test/java/com/havefunwith/TestcontainersTest.java
@@ -1,21 +1,0 @@
-package com.havefunwith;
-
-import org.flywaydb.core.Flyway;
-import org.junit.jupiter.api.Test;
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
-import org.testcontainers.containers.PostgreSQLContainer;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-public class TestcontainersTest extends AbstractTestcontainers {
-
-    @Test
-    void canStartPostgresDB() {
-        assertThat(postgreSQLContainer.isRunning()).isTrue();
-        assertThat(postgreSQLContainer.isCreated()).isTrue();
-    }
-
-}
